Number of Coins Total
I think we should keep the total mined as the biggest 64-bit int because this gives us the most precision and is the tested default.  I think we should think of each pollen as 10^9 atomic units, so there will be 184M total coins.  This matches this analysis: https://docs.google.com/spreadsheets/d/1TAascOuYzjkJh3R2ZVKZp8e6aD5cMWJ6MmXRLd36cR0/edit#gid=0

Note: the COIN const is not used in code rn but is immortalized for reference.

Displayed digits set to 14.  Monero + 2