Emissions curve changes
Changed DIFFICULTY_TARGET to 300 seconds.  This will be the average time between blocks.  I kept it long for now to decrease blockchain size.  If we notice crowded blocks or slow confirmation times we can increase it with a hard fork.  The EMISSION_SPEED_FACTOR was set to match BTC's emission curve (See https://imgur.com/a/LZj4s).  The CRYPTONOTE_BLOCK_GRANTED_FULL_REWARD_ZONE was increased to match the longer blocks.